requirements.txt
imap_tools
fastmcp
python-dotenv
